CLN: new modules conversions and spectra
